golang: extract out "goget" task
